changed caching class to be the RMP DAO, dependent on the API class
